Correcao Avaliacao

+++ src/main/java/com/api/meal4you/repository/SocialLoginRepository.java
@@ -6,7 +6,7 @@ import com.api.meal4you.entity.AdmRestaurante;
 import com.api.meal4you.entity.SocialLogin;
 import com.api.meal4you.entity.Usuario;
 
-public interface SocialLoginRepository extends JpaRepository<SocialLogin, Long> {
+public interface SocialLoginRepository extends JpaRepository<SocialLogin, Integer> {
 
     void deleteByUsuario(Usuario usuario);
